Refactor dummy_country to make it look like france
