+++ openfisca_core/tests/dummy_country/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from dummy_taxbenefitsystem import DummyTaxBenefitSystem
+
+CountryTaxBenefitSystem = DummyTaxBenefitSystem
