+++ openfisca_core/tests/dummy_extension/paris.py
@@ -1,7 +1,7 @@
 from openfisca_core.columns import FloatCol
 from openfisca_core.variables import Variable
 
-from openfisca_core.tests.dummy_country import Famille
+from openfisca_core.tests.dummy_country.entities import Famille
 
 
 class paris_logement_familles(Variable):
