+++ openfisca_core/tests/test_extra_params.py
@@ -5,7 +5,7 @@ from openfisca_core import periods
 from openfisca_core.columns import IntCol, BoolCol
 from openfisca_core.variables import Variable
 from openfisca_core.tests import dummy_country
-from openfisca_core.tests.dummy_country import Individu
+from openfisca_core.tests.dummy_country.entities import Individu
 from openfisca_core.tools import assert_near
 from openfisca_core.base_functions import requested_period_last_value
