+++ openfisca_core/tests/test_opt_out_cache.py
@@ -4,7 +4,7 @@
 from openfisca_core.tests import dummy_country
 from openfisca_core.variables import Variable
 from openfisca_core.columns import IntCol
-from openfisca_core.tests.dummy_country import Individu
+from openfisca_core.tests.dummy_country.entities import Individu
 
 
 class input(Variable):
